requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
python-dotenv==1.0.1
selenium==4.22.0
requests==2.28.1
webdriver-manager==4.0.1